update table migrate

+++ database/migrations/2023_05_16_042625_create_exchanges_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('exchanges', function (Blueprint $table) {
             $table->id();
             $table->integer('dollar');
-            $table->float('khmer');
+            $table->integer('khmer');
             $table->boolean('status');
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrentOnUpdate()->nullable();
